move hyperdome server interactions to seperate file, rename threads to tasks
